refactor(meetings): consolidate schedule + projection into MeetingBriefingsService
The schedule loader and rrule projection were each their own service
with one method, plus controller injection. Three providers for two
methods is more module/DI plumbing than it's worth — fold both into
the existing MeetingBriefingsService (PrismaBase on MeetingBriefing,
which already has the prisma client and now takes S3Service as its
only extra dep).

- loadLatestScheduleForOrg(orgSlug) — replaces MeetingScheduleService
- projectMeetingDates(args) — replaces MeetingProjectionService

Controller drops two constructor params. Module drops two providers.
Tests unchanged (drove through the HTTP routes already).
