+++ src/meetings/services/meetingBriefings.service.ts
@@ -1,7 +1,72 @@
 import { Injectable } from '@nestjs/common'
+import { ExperimentRunStatus, Prisma } from '@prisma/client'
+import { rrulestr } from 'rrule'
+import { formatInTimeZone } from 'date-fns-tz'
 import { createPrismaBase, MODELS } from 'src/prisma/util/prisma.util'
+import { S3Service } from '@/vendors/aws/services/s3.service'
+import {
+  MeetingScheduleArtifact,
+  MeetingScheduleArtifactSchema,
+} from '@goodparty_org/contracts'
+
+const SCHEDULE_EXPERIMENT_TYPE = 'meeting_schedule'
+
+export type ProjectArgs = {
+  schedule: MeetingScheduleArtifact
+  from: Date
+  to: Date
+}
 
 @Injectable()
 export class MeetingBriefingsService extends createPrismaBase(
   MODELS.MeetingBriefing,
-) {}
+) {
+  constructor(private readonly s3: S3Service) {
+    super()
+  }
+
+  async loadLatestScheduleForOrg(
+    organizationSlug: string,
+  ): Promise<MeetingScheduleArtifact | null> {
+    const run = await this.client.experimentRun.findFirst({
+      where: {
+        organizationSlug,
+        experimentType: SCHEDULE_EXPERIMENT_TYPE,
+        status: ExperimentRunStatus.COMPLETED,
+        artifactBucket: { not: null },
+        artifactKey: { not: null },
+      },
+      orderBy: { createdAt: Prisma.SortOrder.desc },
+    })
+    if (!run || !run.artifactBucket || !run.artifactKey) return null
+
+    const raw = await this.s3.getFile(run.artifactBucket, run.artifactKey)
+    if (!raw) return null
+
+    try {
+      const parsed = MeetingScheduleArtifactSchema.safeParse(JSON.parse(raw))
+      return parsed.success ? parsed.data : null
+    } catch {
+      return null
+    }
+  }
+
+  projectMeetingDates({ schedule, from, to }: ProjectArgs): string[] {
+    if (schedule.status === 'not_found') return []
+
+    try {
+      const anchorDate = formatInTimeZone(from, schedule.timezone, 'yyyyMMdd')
+      const anchorTime = schedule.time.replace(':', '') + '00'
+
+      const rule = rrulestr(
+        `DTSTART:${anchorDate}T${anchorTime}\nRRULE:${schedule.rrule}`,
+      )
+
+      return rule
+        .between(from, to, true)
+        .map((d) => formatInTimeZone(d, 'UTC', 'yyyy-MM-dd'))
+    } catch {
+      return []
+    }
+  }
+}
